Finish g2p/p2g tests
