+++ multibody/mpm/sparse_grid.h
@@ -15,7 +15,7 @@
 #include "SPGrid_Page_Map.h"
 
 #include "drake/common/eigen_types.h"
-#include "drake/multibody/mpm/bspline_weights.h"
+#include "drake/multibody/mpm/math.h"
 
 namespace drake {
 namespace multibody {
